New command options and automatic pagination.
Added --merges and --reasons options to show Merge: and Merge*: lines.
(They're now suppressed by default.)

Added a --width argument that currently does nothing. :p
BUT, it's now properly detecting the width in Linux and OSX if it's not provided.

If 'log' detects that STDOUT is a tty, it will automatically pipe its output through 'less'.